moving the .js files to src folder
