+++ src/block-reemit.js
