+++ src/block-ref-rewrite.js
