+++ src/block-tracker-inspector.js
