+++ src/wallet-parity.js
