+++ test/block-ref.js
@@ -4,10 +4,10 @@ const BlockTracker = require('eth-block-tracker')
 const GanacheCore = require('ganache-core')
 const pify = require('pify')
 const EthQuery = require('ethjs-query')
-const createBlockRefMiddleware = require('../block-ref')
-const ScaffoldMiddleware = require('../scaffold')
-const providerFromEngine = require('../providerFromEngine')
-const providerAsMiddleware = require('../providerAsMiddleware')
+const createBlockRefMiddleware = require('../src/block-ref')
+const ScaffoldMiddleware = require('../src/scaffold')
+const providerFromEngine = require('../src/providerFromEngine')
+const providerAsMiddleware = require('../src/providerAsMiddleware')
 const createHitTrackerMiddleware = require('./util/createHitTrackerMiddleware')
 
 test('contructor - no opts', (t) => {
