+++ test/wallet.js
@@ -5,9 +5,9 @@ const EthQuery = require('ethjs-query')
 const GanacheCore = require('ganache-core')
 const pify = require('pify')
 // const providerAsMiddleware = require('../providerAsMiddleware')
-const providerFromEngine = require('../providerFromEngine')
+const providerFromEngine = require('../src/providerFromEngine')
 // const createScaffoldMiddleware = require('../scaffold')
-const createWalletMiddleware = require('../wallet')
+const createWalletMiddleware = require('../src/wallet')
 
 const testAddresses = ['0xbe93f9bacbcffc8ee6663f2647917ed7a20a57bb', '0x1234362ef32bcd26d3dd18ca749378213625ba0b']
 const testUnkownAddress = '0xbadbadbadbadbadbadbadbadbadbadbadbadbad6'
